requirements.txt
fastapi
uvicorn
boto3
databases
alembic
python-dotenv
passlib
bcrypt
python-jose
pydantic
pydantic[email]
psycopg2-binary
mangum
pytest
httpx